refactor: Move collectUniqueAndPairedCaveats to utils.ts
