+++ packages/permission-controller/src/index.ts
@@ -2,6 +2,12 @@ export * from './Caveat';
 export * from './errors';
 export * from './Permission';
 export * from './PermissionController';
-export * from './utils';
+export type {
+  ExtractSpecifications,
+  HandlerMiddlewareFunction,
+  HookNames,
+  PermittedHandlerExport,
+} from './utils';
+export { MethodNames } from './utils';
 export * as permissionRpcMethods from './rpc-methods';
 export * from './SubjectMetadataController';
